+++ src/main/java/nadun_blog/repo/RoleRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import nadun_blog.model.Role;
 
 @Repository
-public interface RoleRepo extends JpaRepository<Role, Long> {
+public interface RoleRepo extends JpaRepository<Role, Integer> {
 
 }
